app control of a repeating crossbow
The Chinese continuous crossbow, commonly referred to as the repeating crossbow, stands as an ingenious ancient weapon with origins in China. Its design was tailored for manual operation, enabling rapid firing of multiple shots. The primary goal of this project is to achieve remote control over the crossbow through an Android app, utilizing Bluetooth communication.